Added class declarations for cfg

+++ src/edu/mit/compilers/lowir/instructions/Enter.java
@@ -1,6 +1,6 @@
 package edu.mit.compilers.lowir.instructions;
 
 public class Enter extends Instruction {
-    public Enter(uint _numVariables) {
+    public Enter(long _numVariables) {
     }
 }
